🩹 [Patch]: Allow external continue-on-error, and add status in action (#11)
## Description

This pull request includes several updates to the GitHub Actions
workflow configuration and test files to improve the handling and
reporting of test outcomes and conclusions.

Enhancements to GitHub Actions workflow:

* `.github/workflows/Action-Test.yml`
* Added outputs for `outcome` and `conclusion` to multiple jobs,
including `ActionTest1Simple`, `ActionTest1SimpleFailure`,
`ActionTest2Standard`, and `ActionTest3Advanced`.
* Introduced a new job `CatchJob` to aggregate and display the results
of all test jobs in a table format, and to check if all jobs passed
their expected outcomes and conclusions.

Updates to individual test cases:

* `action.yml`: Added a new step to check the `Passed` status of tests
and exit with an error if the tests did not pass, which in turn makes
the caller be able to distinguish if the job failed and allows them to
act accordingly. We are using this mechanism in the `CatchJob` step of
the `Action-Test` workflow to aggregate all status.
* `tests/1-Simple-Failure/Failure.Tests.ps1`: Added a new test case to
intentionally fail by asserting that `true` should be `false`, which is
required for testing failure scenarios.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas