🌟 [Major]: Output cleanup + Fix test name expansion in summary (#12)
## Description

🌟 This pull request is a breaking change as it changes how you would
determine if the tests are successful or not. Previously this was done
with a `Passed` output, but is now just relying on the `outcome` and
`conclusion` properties of the test step. Together with
`continue-on-error` it creates a full solution for how to work with
failing tests. Don't use `continue-on-error` if you want the workflow to
stop, then and there, or use it with `conclusion` to gather tests and
make a decision based on more criteria.

### Fixing the summary

This pull request started as an attempt to just fix expansion of test
items in the Step Summary, but became more of an overhaul of the
functionality. This effort resulted in the creation of the `Markdown`
module that creates a DSL for writing markdown documents. The function
to traverse and create the summary structure was also heavily simplified
by following the hierarchy of objects in a `Pester.Run` class, resulting
in a recursive function that will also make it easier to add
functionality to it later.

- Fixes #9 

This could have been made simpler by just having "Children" as a
property of every object, but its called different and pretty much
achieves the same. `Invoke-Pester` returns a `Pester.Run`, a run has a
list of `Pester.Container` on the property `Container`. Containers have
`Pester.Block` on a property called `Blocks`. Blocks have
`Pester.Blocks` or `Pester.Test` objects on the property `Order`.

```mermaid
graph LR
    Run --> Container
    Container --> Blocks
    Blocks --> Order
    Order --> Test
    Order --> Blocks
```

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas