🩹 [Patch]: Add basic `GitHub-Script` inputs (#14)
## Description

This pull request includes updates to the `README.md` and `action.yml`
files to add new input options for the `GitHub-Script` action.

Updates to input options:

* `README.md`:
* Modified the description for the `Debug` option and added new options
for `Verbose`, `Version`, and `Prerelease`.
* `action.yml`:
* Updated the `Debug` option description and added new input options
`Verbose`, `Version`, and `Prerelease` with their respective
descriptions and default values.
* Added the new input options `Debug`, `Prerelease`, `Verbose`, and
`Version` to the `runs` configuration to ensure they are passed
correctly to the script.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas