🩹 [Patch]: Refactor to run `Invoke-Pester` in a clean session (#15)
## Description

This pull request includes significant changes to the `action.yml` and
`scripts/main.ps1` files to improve the structure and functionality of
the Pester testing process, including reorganizing steps, modifying
script invocations, and simplifying the code for better readability and
maintainability.
All of this to attempt to run Invoke-Pester in a relatively clean
session.

### Changes to `action.yml`:

* Added a new step `Invoke-Pester (pre)` for gathering tests and
building configuration. Configs are now exported to disk and read in the
following step. -> runs `scripts/pre.ps1`
* Changed the test step to run in a plain `pwsh` session. -> runs
`scripts/main.ps1`

### Added `scripts/pre.ps1`:
* Moved the functionality of gathering inputs and generating the
configuration and containers from the `main.ps1` file. This means we can
reduce the number of prereqs that are needed in the session that will
run `Invoke-Pester`.

### Changes to `scripts/main.ps1`:

* Replaced `LogGroup` blocks with GitHub Actions logging commands
(`::group::` and `::endgroup::`) to avoid loading the GitHub module.
* Simplified the installation and import of necessary modules.
* Removed the detailed manual loading of inputs and configurations,
replacing it with ann import of the saved file
(`Invoke-Pester.Configuration.ps1`) from the pre-step.
* Updated the container handling to dynamically load and process
container files.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas