🪲 [Fix]: Fix default path for Pester, its workspace now (#18)
## Description

This pull request includes changes to the `action.yml` and
`scripts/init.ps1` files to improve configuration handling and
initialization scripts. The most important changes include the removal
of a default value in the `action.yml` file and the reordering of
container initialization steps in the `scripts/init.ps1` file.

Changes to configuration handling:

*
[`action.yml`](diffhunk://#diff-1243c5424efaaa19bd8e813c5e6f6da46316e63761421b3e5f5c8ced9a36e6b6L13):
Removed the default value for the `Path` input to allow for more
flexible test path configurations.

Changes to initialization scripts:

*
[`scripts/init.ps1`](diffhunk://#diff-f47ceebe9ade2bb55cede031de8267e9c87b09336a93fcd557c02c1f488554b6L179):
Moved the reinitialization of the `$configuration.Run.Container` array
to occur after the existing containers are processed, ensuring that the
array is properly reset.
[[1]](diffhunk://#diff-f47ceebe9ade2bb55cede031de8267e9c87b09336a93fcd557c02c1f488554b6L179)
[[2]](diffhunk://#diff-f47ceebe9ade2bb55cede031de8267e9c87b09336a93fcd557c02c1f488554b6R202)

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas