🩹 [Patch]: Update path resolution logic to support recursive search (#26)
## Description

This pull request includes several changes to improve the configuration
handling and directory management in the testing scripts and workflows,
including updates to the workflow configuration, enhancements to the
script logic for handling paths and directories, and improvements to the
configuration export process.

### Script Enhancements

* `scripts/Helpers.psm1`:
* Added the `-Recurse` flag to `Get-ChildItem` to search for
configuration files recursively in directories.
* `scripts/init.ps1`:
* Improved the logic for handling provided paths, ensuring proper
directory and file validation.
* Enhanced the handling of boolean input values by explicitly checking
and converting string representations of boolean values.
* Updated the configuration export process to use a temporary directory
for storing configuration and container files, ensuring cleaner and more
organized output.
* `scripts/main.ps1`:
* Modified the script to import the configuration from the temporary
directory and handle missing configuration files more gracefully.

### Testing changes

* `.github/workflows/Action-Test.yml`:
* Replaced `Path` with `WorkingDirectory` in multiple job steps to
better specify the working directory for tests.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas