🌟 [Breaking change]: Cleanup `StepSummary` inputs (#36)
## Description

This pull request includes several changes to improve the configuration
and functionality of the GitHub step summary in the Pester testing
framework. The most important changes include the consolidation of step
summary controls into a single mode parameter, updates to default
values, and improvements to the logic for generating summaries.

Changes to step summary controls:

* 🌟 Consolidated `StepSummary_Enabled` and `StepSummary_ShowTests` into
a single `StepSummary_Mode` parameter, which controls which tests to
show in the GitHub step summary. The allowed values are:
  * "Full" (show all tests)
  * "Failed" (only failed tests)
  * "None" (disable step summary).

Updates to default values:

* Changed the default value of `StepSummary_ShowTestOverview` to `false`
in both `action.yml` and `scripts/Helpers.psm1`.

Improvements to summary generation logic:

* Updated the logic in `scripts/exec.ps1` to only generate a step
summary if the `StepSummary_Mode` is not set to 'None' and at least one
component (test overview or configuration) is configured to be
displayed.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [x] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas