🪲 [Fix]: Add tests for `Debug-` and `VerbosePreference` (#45)
## Description

This pull request introduces debugging and verbose logging enhancements
to the testing workflow and scripts. The changes are primarily focused
on improving the visibility and traceability of the script executions by
adding debug and verbose options.

Enhancements to logging:

*
[`.github/workflows/Action-Test.yml`](diffhunk://#diff-a12ae5c885b0673c0ff6f70c2670886907590d624626e07da4c52e01aeaf56a4R39-R40):
Added `Debug` and `Verbose` options to the `WorkingDirectory`
configuration to enable detailed logging during the action test
execution.
*
[`scripts/exec.ps1`](diffhunk://#diff-9565c9ed6189efecf687222ac63d0ec9702c9708f75a02f7d9d9de9f86ae37a0R8-R9):
Introduced `$DebugPreference` and updated `$VerbosePreference` to
control the logging behavior based on environment variables, ensuring
that debug and verbose messages are handled appropriately.
*
[`tests/1-Simple/Emoji/Emoji.ps1`](diffhunk://#diff-d612dc40477865ab6616b2aa71e09ab79aa457cdf2f631a8496c7998b62cf820R23-R24):
Added `Write-Verbose` and `Write-Debug` statements to the `Get-Emoji`
function to log the process of retrieving emojis, aiding in debugging
and traceability.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [x] 🪲 [Fix]
- [ ] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas